+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 5.2.0 / wip
 ==================
 
+**Requires referencedata:15.3.0 or later**
+
+Improvements:
+* [OE-86](https://openlmis.atlassian.net/browse/OE-86): Added /api/public/stockCardSummaries endpoint which is an
+ an equivalent of /api/v2/stockCardSummaries service tailored for external integrations
+
 5.1.12 / 19.11.2024
 ==================